@@ -1,4 +1,5 @@
-import '../model/click_count.dart';
+// Project imports:
+import 'entity/click_count.dart';
 
 abstract class CounterRepository {
   Future<ClickCount> loadCounter();
